Merge in ADGUARD-IOS/adguard-ios from enhance/AG-21154 to version/v4.3

* commit '468a9e3f6f500281c22fbdc4bb7c74fcc54cf458':
  🛠 Chore: Bump build version
  🛠 Chore: #upgrade_deps converterlib from:v2.0.38 to:v2.0.39

+++ AdGuardSDK/AdGuardSDK/Safari/SafariProtection/SafariConfiguration.swift
@@ -21,7 +21,7 @@ import SharedAdGuardSDK
 // MARK: - ConfigurationProtocol
 
 public protocol SafariConfigurationProtocol: ConfigurationProtocol {
-    var iosVersion: Int { get set } // iOS version is used in SafariConverter lib, it can be obtained with UIDevice.current.systemVersion
+    var iosVersion: Double { get set } // iOS version is used in SafariConverter lib, it can be obtained with UIDevice.current.systemVersion
     var safariProtectionEnabled: Bool { get set }
     var advancedBlockingIsEnabled: Bool { get set } // Feature is available since iOS 15
     var blocklistIsEnabled: Bool { get set }
@@ -38,7 +38,7 @@ public protocol SafariConfigurationProtocol: ConfigurationProtocol {
 // MARK: - Configuration
 
 public final class SafariConfiguration: SafariConfigurationProtocol {
-    public var iosVersion: Int
+    public var iosVersion: Double
     public var currentLocale: Locale
     public var proStatus: Bool
     public var safariProtectionEnabled: Bool
@@ -53,7 +53,7 @@ public final class SafariConfiguration: SafariConfigurationProtocol {
     public var cid: String
 
     public init(
-        iosVersion: Int,
+        iosVersion: Double,
         currentLocale: Locale,
         proStatus: Bool,
         safariProtectionEnabled: Bool,
